1st exercise completed, to identify polygon by number of sides
Read the number of sides the polygon has and display which polygon it is.
a) If the polygon has 3 sides it is a triangle
b) If the polygon has 4 sides it is a quadrilateral
c) If the polygon has 5 sides it is pentagon, etc.

To do this, a Python Tuple has been created, with names of polygons 1-12.
Then, through an input function, the user specifies the number of sides.
For example, 5 sides = item number 5 in the tuple, which is a pentagon.